libft compiles also

+++ lib/libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: vviterbo <vviterbo@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/01 12:33:18 by vviterbo          #+#    #+#             */
-/*   Updated: 2025/06/12 01:35:05 by vviterbo         ###   ########.fr       */
+/*   Updated: 2025/06/16 11:28:57 by vviterbo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -136,10 +136,10 @@ t_coor		ft_coorsubstr(t_coor v1, t_coor v2);
 double		*ft_detmat(t_mat mat);
 double		ft_dotprod(t_coor v1, t_coor v2);
 void		ft_free_mat(t_mat *mat);
-float		*ft_getcofac(t_mat mat, int m, int n);
-t_mat		*ft_getsubmat(t_mat mat, int m, int n);
+double		*ft_getcofac(t_mat mat, int m, int n);
+t_mat		*ft_getsubmat(t_mat	mat, size_t m, size_t n);
 t_coor		ft_initcoor(double x, double y, double z);
-t_mat		*ft_initmat(int m, int n);
+t_mat		*ft_initmat(size_t m, size_t n);
 double		ft_max(double a, double b);
 double		ft_min(double a, double b);
 double		ft_norm(t_coor coor);
